Make enemies' waves and appear key
Set a manager to which the enemies must be added and a minimum number
before the next wave must be set, and the golden key too. After some
enemies are eliminated, the second enemies' wave appear and, after those
are eliminated, the key appears in the position of the last enemy
killed.